perf: split firebase init
