+++ src/lib/firebase/auth.ts
@@ -1,0 +1,5 @@
+import { getAuth } from "firebase/auth"
+
+import { app } from "./init"
+
+export const auth = getAuth(app)
