+++ src/lib/firebase/db.ts
@@ -1,0 +1,5 @@
+import { getFirestore } from "firebase/firestore"
+
+import { app } from "./init"
+
+export const db = getFirestore(app)
