+++ src/lib/firebase/storage.ts
@@ -1,0 +1,5 @@
+import { getStorage } from "firebase/storage"
+
+import { app } from "./init"
+
+export const storage = getStorage(app)
